pyproject.toml
[project]
name = "deepscholar"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "zhipuai>=2.1.0",
    "sniffio>=1.3.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "bus", "mcp_servers"]
